Move utxo persistent logic testing to utxo_storage.rs for more isolation between logic and storage
