+++ blockchain_storage/Cargo.toml
@@ -17,7 +17,6 @@ thiserror = "1.0"
 [dev-dependencies]
 mockall = "0.11"
 rand = "0.8.4"
-iter_tools = "0.1.0"
 
 [features]
 mock = [ 'mockall' ]
