+++ utxo/Cargo.toml
@@ -9,3 +9,4 @@ edition = "2021"
 common = { path = '../common' }
 parity-scale-codec = { version = "2.3.1", features = ["chain-error"] }
 rand = "0.8.4"
+iter_tools = "0.1.0"
